# gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 百舸API网关
  description: |
    统一百舸各模块API规范和数据结构的中间层网关服务。

    所有接口采用RPC风格，通过 `?action=XXX` 参数区分操作。
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000/v1
    description: 本地开发服务器
...
paths:
...
  /aihc/models/DescribeModels:
    get:
      tags:
        - Models
      summary: Models - DescribeModels
      operationId: describeModels
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: modelId
          in: query
          schema:
            type: string
          description: 模型ID
        - name: keyword
          in: query
          schema:
            type: string
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelListResponse"
  /aihc/models/DescribeModel:
    get:
      tags:
        - Models
      summary: Models - DescribeModel
      operationId: describeModel
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: modelId
          in: query
          schema:
            type: string
          description: 模型ID
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelDetail"
  /aihc/models/DescribeModelVersions:
    get:
      tags:
        - Models
      summary: Models - DescribeModelVersions
      operationId: describeModelVersions
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: modelId
          in: query
          schema:
            type: string
          description: 模型ID
        - name: keyword
          in: query
          schema:
            type: string
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelVersionListResponse"
  /aihc/models/CreateModel:
    post:
      tags:
        - Models
      summary: Models - CreateModel
      operationId: createModel
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModelRequest"
      responses:
        "200":
          description: 成功
        "201":
          description: 创建成功
...
  /aihc/models/DeleteModel:
    post:
      tags:
        - Models
      summary: Models - DeleteModel
      operationId: deleteModel
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteModelRequest"
      responses:
        "200":
          description: 成功
        "201":
          description: 创建成功
...
components:
  parameters:
    RegionParam:
      name: region
      in: header
      required: true
      description: 区域代码
      schema:
        type: string
        enum:
          - bj
          - gz
          - su
          - bd
          - fwh
          - yq
        default: bd
    AkParam:
      name: ak
      in: header
      required: true
      description: 百度云 Access Key
      schema:
        type: string
    SkParam:
      name: sk
      in: header
      required: true
      description: 百度云 Secret Key
      schema:
        type: string
  schemas:
    Pagination:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        totalCount:
          type: integer
...
    CreateModelRequest:
      type: object
      required:
        - name
        - modelFormat
      properties:
        name:
          type: string
        modelFormat:
          type: string
        description:
          type: string
        owner:
          type: string
        visibilityScope:
          type: string
          enum:
            - all
            - owner
            - group
    DeleteModelRequest:
      type: object
      required:
        - modelId
      properties:
        modelId:
          type: string
    ModelListResponse:
      type: object
      properties:
        models:
          type: array
          items:
            type: object
        pagination:
          $ref: "#/components/schemas/Pagination"
    ModelDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        modelFormat:
          type: string
    ModelVersionListResponse:
      type: object
      properties:
        versions:
          type: array
          items:
            type: object
        pagination:
          $ref: "#/components/schemas/Pagination"
...